# Optimize //src:bazel-dev / package-bazel.sh to reduce incremental build times by 60%