@meisterT 2020-01-15
Do we run on CI with the tests with -c opt?