@meisterT 2020-01-15
Thanks for checking, Philipp.

@jin either add `-c opt` then to CI or use the `bazel-dev` target as we do on Windows. I want to run the tests on CI with the binary that we are releasing.