@jin 2020-01-15
@philwo do we build release builds with `-c opt`? I can't find this configured anywhere in our release scripts / source tree.