@meisterT 2020-01-15
Yes, I added this in https://github.com/bazelbuild/bazel/commit/cc004fbe7a7cadfe89f04ab644fcfc732fe6b2e7