@jin 2020-01-15
@meisterT Sounds good, I added it to bazel-dev instead. I've also updated the docs to make it more prominent that `//src:bazel-dev` is the dev build. What do you think?