@jin 2020-01-15
(FYI @katre for bazel-dev.sh changes)